Improve handling of query server errors
