+++ ui/src/main/webapp/app.js
@@ -1,13 +1,13 @@
 import React from 'react'
 import Home from './components/Home'
 import Navbar from './components/Navbar'
-import SitesPage from './components/sites/SitesPage'
 import { HashRouter, Route, Switch, Link } from 'react-router-dom'
 import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles'
 import client from './client'
 import { ApolloProvider } from 'react-apollo'
 import { SnackbarProvider } from 'notistack'
 import { Grid, Typography } from '@material-ui/core'
+import SitesContainer from './components/sites/SitesContainer'
 
 const theme = createMuiTheme({
   // see: https://material-ui.com/style/typography/#migration-to-typography-v2
@@ -50,7 +50,7 @@ const App = () => {
               <div style={{ margin: 10 }}>
                 <Switch>
                   <Route exact path='/' component={Home} />
-                  <Route exact path='/nodes' component={SitesPage} />
+                  <Route exact path='/nodes' component={SitesContainer} />
                   <Route
                     render={function() {
                       return <NotFoundPage />
